Remove helper module as it is redundant
